Fix some of the StrictMode violations
Mostly it's disk reads/writes. I fixed those for AccountManager which
are reported at startup.

Conscrypt reads own version on startup which reads from disk multiple
times. There's no solution for it right now.

SharedPreferences which are used in BaseActivity also read from disk
and pretty early but it shouldn't be a problem.